% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Split a soccer pitch into user defined zones}
\usage{
create_grid(
  pitch_length = pitch_length_init,
  pitch_width = pitch_width_init,
  n_columns = 6,
  n_rows = 5
)
}
\arguments{
\item{pitch_length}{Length of the pitch}

\item{pitch_width}{Width of the pitch}

\item{n_columns}{Number of columns}

\item{n_rows}{Number of rows}
}
\value{
A tibble with zone definitions
}
\description{
Split a soccer pitch into user defined zones
}
\examples{
my_grid <- create_grid(120, 75, 5, 6)
}
