% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_num_name.R
\name{display_num_name}
\alias{display_num_name}
\title{Diplay Number and Last Name}
\usage{
display_num_name(number, name)
}
\arguments{
\item{number}{The player's jersey number}

\item{name}{The player's name to display}
}
\description{
This function displays name in the form #number name
}
