% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pitch_org.R
\name{draw_pitch_orig}
\alias{draw_pitch_orig}
\title{Create a pitch plot ready for Statsbomb data}
\usage{
draw_pitch_orig(
  pitch_length = 105,
  pitch_width = 68,
  pitch_units = "meters",
  pitch_layout = "horizontal",
  pitch_section = "full",
  grass_colour = "#ffffff",
  line_colour = "#A9A9A9",
  background_colour = "#ffffff",
  goal_colour = "#000000",
  goaltype = "line",
  middlethird = FALSE,
  BasicFeatures = FALSE,
  overlay = FALSE,
  arcs = TRUE,
  padding = 5
)
}
\arguments{
\item{pitch_length}{Length of the pitch. Always the long side. Default 105}

\item{pitch_width}{Width of the pitch. Default 68}

\item{pitch_units}{Units of measure in "meters" or "yards". Default "meters"}

\item{pitch_layout}{Plot the pitch vertical or horizontal. Default "horizontal"}

\item{pitch_section}{Full or subsection of field. "full", "att_half". Default "full"}

\item{grass_colour}{A string as a colour by word or #HEX code}

\item{line_colour}{A string as a colour by word or #HEX code}

\item{background_colour}{A string as a colour by word or #HEX code}

\item{goal_colour}{A string as a colour by word or #HEX code}

\item{goaltype}{String, either "line", "box" or "barcanumbers". Default "line"}

\item{middlethird}{Boolean. Default FALSE}

\item{BasicFeatures}{Boolean. Default FALSE}

\item{overlay}{Draw lines on top of the field. "JdeP" "grid_30" or FALSE}

\item{arcs}{Boolean. Default TRUE}

\item{padding}{Numeric. Default = 5}
}
\value{
A plot of a pitch
}
\description{
Creating a pitch plot using ggplot2 that is ready to plot actions on top. Numerous options allow customisation.
}
