% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_sim_match_points.R
\docType{data}
\name{example_df}
\alias{example_df}
\title{Simulate the number of goals scored given a vector of shot xG values for two teams}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.
}
\usage{
example_df
}
\arguments{
\item{team_b_shots_xg}{}
}
\description{
Simulate the number of goals scored given a vector of shot xG values for two teams
}
\keyword{datasets}
