% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_within_timeframe.R
\name{filter_within_timeframe}
\alias{filter_within_timeframe}
\title{Find and filter all events within a certain time of a specific event
For example, find all events 5 seconds prior to and including a shot
Right now only looks works for seconds before (lead), need to have a solution that looks after and between two times}
\usage{
filter_within_timeframe(.data, condition, time, lead_time = 0, lag_time = 0)
}
\arguments{
\item{.data}{The data frame}

\item{condition}{The event to find in dplyr fashion. Such as event == 'A'}

\item{time}{The column/data that contains the timestamp}

\item{lead_time}{The number of seconds before the event occurs}

\item{lag_time}{The number of seconds after an event occurs (needs to be implemented)}

\item{time_filter}{The number of seconds before the event occurs}
}
\value{
A plot of a pitch
}
\description{
Find and filter all events within a certain time of a specific event
For example, find all events 5 seconds prior to and including a shot
Right now only looks works for seconds before (lead), need to have a solution that looks after and between two times
}
