% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plot_bar_range.R
\name{gt_plot_bar_range}
\alias{gt_plot_bar_range}
\title{After gtExtras::gt_plot_bar_stack. The gt_plot_bar_range takes an existing gt_tbl object and converts the existing values into a stacked barchart.}
\usage{
gt_plot_bar_range(gt_object, column)
}
\arguments{
\item{gt_object}{An existing gt table object of class gt_tbl}

\item{column}{The column wherein the  stacked barchart should replace existing data. Note that the data must be represented as a list of three numeric values and an optional color value ahead of time.}

\item{team_a_name}{name of team A}

\item{team_b_name}{name of team B}

\item{n_sim}{number of simulations to run}

\item{seed}{seed for reproducibility}
}
\value{
A list of two data frames: 1) win-win-draw probabilities and 2) result probabilities
}
\description{
After gtExtras::gt_plot_bar_stack. The gt_plot_bar_range takes an existing gt_tbl object and converts the existing values into a stacked barchart.
}
\examples{
\dontrun{
ex_df <- tibble(
x = c("load", "sumad", "spd", "hsrp", "tdc"),
list_data = list(c(75, 50, 75, "purple"), c(150, 50, 0, "blue"), c(0, 25, 175, "red"),
                 c(50, 100, 50, "orange"), c(50, 50, 100, "green"))
)

ex_tab <- ex_df \%>\%
  gt() \%>\%
  gt_plot_bar_range(column = list_data)

ex_tab
}
}
