% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_90.R
\name{per_90}
\alias{per_90}
\title{Per 90 Converstion}
\usage{
per_90(stat, minutes_played)
}
\arguments{
\item{stat}{The stat to convert to a per 90 metric}

\item{minutes_played}{The total number of minutes played}
}
\description{
This function converts a raw stat to a per 90 metric
}
