% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch_setup.R
\name{pitch_setup}
\alias{pitch_setup}
\title{Initialize dimensions for use elsewhere}
\usage{
pitch_setup(provider, length, width)
}
\arguments{
\item{provider}{Shortcut to add dimensions of provider}

\item{pitch_length_init}{Length of the pitch.}

\item{pitch_width_init}{Width of the pitch.}

\item{pitch_units_init}{Units of measurement.}
}
\value{
Variables to be used in the future
}
\description{
Initialize dimensions for use elsewhere
}
