% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_logo.R
\name{plot_add_logo}
\alias{plot_add_logo}
\title{Add logo to a plot}
\usage{
plot_add_logo(
  plot_path,
  logo_path,
  logo_position,
  logo_scale = 10,
  logo_padding = 0.01
)
}
\arguments{
\item{plot_path}{Location of plot image to add logo to}

\item{logo_path}{Location of logo image to add to plot}

\item{logo_position}{Position of logo on plot. Options are "top right", "top left", "bottom right", "bottom left"}

\item{logo_scale}{Scale of logo relative to plot width. Default is 10\% of plot width}

\item{logo_padding}{Padding between logo and plot edge. Default is 1\% of plot width}
}
\value{
An image file
}
\description{
Add logo to a plot
}
