% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar_man.R
\name{plot_radar_man}
\alias{plot_radar_man}
\title{Creates a pizza plot with no preset groups or colors. The data frame must have the
following columns named exactly as shown: player, statistic, value, percentile, category, and color. The plot will
follow the order of the data frame.}
\usage{
plot_radar_man(
  df,
  title,
  subtitle = "",
  caption = "",
  pos_group = "",
  theme_color
)
}
\arguments{
\item{df}{A data frame with six columns: player, statistic, value, percentile, category and color. \code{percentile} is expected to be a number between 0 and 99.}

\item{title}{Text to appear as the title}

\item{subtitle}{Text to appear in the subtile. Use paste(..., sep = "\n") to create multiple lines.}

\item{pos_group}{The position group to use for the average line text. Not required}

\item{theme_color}{The theme color to use: light or dark. Default is light.}
}
\value{
A ggplot object
}
\description{
Creates a pizza plot with no preset groups or colors. The data frame must have the
following columns named exactly as shown: player, statistic, value, percentile, category, and color. The plot will
follow the order of the data frame.
}
\examples{
scaled_data <- data.frame(player = rep("Player 1", 12),
                          statistic = c("Stat 1", "Stat 2", "Stat 3", "Stat 4", "Stat 5", "Stat 6",
                                        "Stat 7", "Stat 8", "Stat 9", "Stat 10", "Stat 11", "Stat 12"),
                          value = c(12, 4, 9, 10, 5, 23, 15, 9, 3, 8, .79, 3),
                          percentile = c(60, 60, 70, 70, 90, 80, 40, 30, 50, 90, 99, 80),
                          category = rep(c("Attack", "Possession", "Defending", "Set Pieces"), each = 3),
                          color = rep(c("#008000", "#0000FF", "#FF0000", "#FFA500") ,each = 3))

p <- plot_radar_man(scaled_data,
                    title = "Player A",
                    subtitle = paste("This Season | Compared to DMs","All Data per 90", sep = "\n"),
                    pos_group = "DM",
                    theme_color = "dark")
p
}
