% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_prob.R
\name{plot_score_prob}
\alias{plot_score_prob}
\title{Plots matrix of score probabilities}
\usage{
plot_score_prob(xg_win_prob_df)
}
\arguments{
\item{xg_win_prob_df}{A dataframe created with xg_win_prob()}
}
\description{
Plots matrix of score probabilities
}
