% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_y.R
\name{transform_y}
\alias{transform_y}
\title{Transform wyscout y coordinate percent to y coordinate (yards)}
\usage{
transform_y(y_coordinate, att_direction = "E")
}
\arguments{
\item{y_coordinate}{y coordinate in percent.}

\item{att_direction}{Attacking direction of the team. Default is "E" (East).}
}
\value{
Transformed y coordinate
}
\description{
Transform wyscout y coordinate percent to y coordinate (yards)
}
