% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_sim_goal.R
\name{xg_sim_goal}
\alias{xg_sim_goal}
\title{xg_sim_goal}
\usage{
xg_sim_goal(shot_xg_list)
}
\arguments{
\item{shot_xg_list}{List of shot xG values}
}
\value{
The simulated number of goals scored
}
\description{
Simulate the number of goals scored given a vector of shot xG values
}
\examples{
set.seed(123)
# Create a data frame of shots
shots <- data.frame(shot = seq(1:10), xG = runif(n = 10, max = 1))

# Simulate the number of goals scored
shots |> pull(xG) |> xg_sim_goal()

}
