% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_simulate_match.R
\name{xg_sim_match}
\alias{xg_sim_match}
\title{Simulate the number of goals scored given a vector of shot xG values for two teams}
\usage{
xg_sim_match(team_a_shots_xg, team_b_shots_xg)
}
\arguments{
\item{team_a_shots_xg}{A list containing the xG values for each shot taken by team A}

\item{team_b_shots_xg}{A list containing the xG values for each shot taken by team B}
}
\description{
Simulate the number of goals scored given a vector of shot xG values for two teams
}
