% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_logo.R
\name{plot_add_logo}
\alias{plot_add_logo}
\title{Add logo to a plot}
\usage{
plot_add_logo(
  plot_path,
  logo_path,
  logo_position,
  logo_scale = 10,
  logo_padding = 0.01
)
}
\arguments{
\item{logo_padding}{}
}
\value{
A
}
\description{
Add logo to a plot
}
